# Hash an actual credential for verifyNoCredential so the timing is input-dependent