@olorin 2016-05-02
Sorry, put this on the wrong PR.

No significant change in the benchmarks.

old:

```
∃ ~/tmp/tinfoil-bench-master kdf
benchmarking kdf/scrypt/hashCredential/defaultParams
time                 1.804 s    (1.653 s .. 1.935 s)
                     0.999 R²   (0.997 R² .. 1.000 R²)
mean                 1.794 s    (1.772 s .. 1.812 s)
std dev              28.73 ms   (0.0 s .. 31.35 ms)
variance introduced by outliers: 19% (moderately inflated)

benchmarking kdf/scrypt/verifyNoCredential/defaultParams
time                 1.676 s    (1.559 s .. 1.847 s)
                     0.998 R²   (0.998 R² .. 1.000 R²)
mean                 1.872 s    (1.785 s .. 1.935 s)
std dev              94.59 ms   (0.0 s .. 108.5 ms)
variance introduced by outliers: 19% (moderately inflated)
```

new:

```
∃ ~/tmp/tinfoil-bench-no-credential kdf
benchmarking kdf/scrypt/hashCredential/defaultParams
time                 1.781 s    (1.777 s .. 1.783 s)
                     1.000 R²   (1.000 R² .. NaN R²)
mean                 1.772 s    (1.768 s .. 1.774 s)
std dev              3.777 ms   (0.0 s .. 4.261 ms)
variance introduced by outliers: 19% (moderately inflated)

benchmarking kdf/scrypt/verifyNoCredential/defaultParams
time                 1.781 s    (1.555 s .. 2.069 s)
                     0.997 R²   (0.990 R² .. 1.000 R²)
mean                 1.845 s    (1.783 s .. 1.883 s)
std dev              57.59 ms   (0.0 s .. 65.79 ms)
variance introduced by outliers: 19% (moderately inflated)
```