@markhibberd 2016-05-02
👍 But I would like the salt change if there are no objections (don't mind if a different change to this though).